% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_priors.R
\name{calculate_priors}
\alias{calculate_priors}
\title{Function to produce a basic Bayesian analysis of artist nationality
probabilities for each quartile in the next Hottest 100 Countdown. The
purpose of this is to use flat priors in order to fit a model and assess
the distribution of coefficient values before using outputs as an informative
prior in a predictive model}
\usage{
calculate_priors(timescale = c("Last Decade", "All Time"))
}
\arguments{
\item{timescale}{a value of either ["Last Decade", "All Time"] indicating the timescale to calculate over}
}
\value{
a dataframe of Bayesian Stan model outputs with 4 columns [mean, sd, lower, upper] where "lower" and "upper" form the 95% credible interval
}
\description{
Function to produce a basic Bayesian analysis of artist nationality
probabilities for each quartile in the next Hottest 100 Countdown. The
purpose of this is to use flat priors in order to fit a model and assess
the distribution of coefficient values before using outputs as an informative
prior in a predictive model
}
\author{
Trent Henderson
}
