% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_countdown_analysis.R
\name{do_countdown_analysis}
\alias{do_countdown_analysis}
\title{Function to produce a small analysis document for a given year's Hottest 100 Countdown
#' This function should ideally take the dataframe output straight from hotteR::get_countdowns()}
\usage{
do_countdown_analysis(data)
}
\arguments{
\item{data}{The dataframe of Hottest 100 results to analyse}
}
\description{
Function to produce a small analysis document for a given year's Hottest 100 Countdown
#' This function should ideally take the dataframe output straight from hotteR::get_countdowns()
}
\author{
Trent Henderson
}
