% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_plays_analysis.R
\name{do_plays_analysis}
\alias{do_plays_analysis}
\title{Function to produce a small analysis document for a given selection of song plays based on tweets
This function should ideally take the dataframe output straight from hotteR::get_plays()}
\usage{
do_plays_analysis(data)
}
\arguments{
\item{data}{The dataframe of triplejplays data to analyse}
}
\value{
Returns an object of class ggplot which contains a matrix of plots
}
\description{
Function to produce a small analysis document for a given selection of song plays based on tweets
This function should ideally take the dataframe output straight from hotteR::get_plays()
}
\examples{
\dontrun{
d <- get_plays(year = 2020)
do_plays_analysis(d)
}

}
\author{
Trent Henderson
}
