% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Sets default geom colours to be compliant with hotteR palette}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{the data library name}

\item{pkgname}{the package name}
}
\value{
default theme for ggplot
}
\description{
Sets default geom colours to be compliant with hotteR palette
}
