% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plays.R
\name{get_plays}
\alias{get_plays}
\title{Function to pull tweets from the 'triplejplays' Twitter account where plays are stored}
\usage{
get_plays(year = 2020)
}
\arguments{
\item{year}{The year of interest to pull tweets from - only current year is possible due to API limitations}
}
\description{
Function to pull tweets from the 'triplejplays' Twitter account where plays are stored
}
\author{
Trent Henderson
}
