% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_countdowns}
\alias{historical_countdowns}
\title{All historical Triple J Hottest 100 Countdown results}
\format{
A dataframe with 3500 rows and 5 variables:
\describe{
  \item{rank}{Position in the Hottest 100 Countdown of a given year (1--100)}
  \item{artist}{Name of the artist}
  \item{song}{Name of the song}
  \item{year}{Year of the countdown (2020 (Of The Decade), 2019, 2018, 2017,2016, 2015, 2014, 2013 (Twenty Years), 2013, 2012, 2011 (Australian Albums), 2011,2010, 2009 (All-time), 2009, 2008,2007, 2006, 2005, 2004, 2003, 2002, 2001, 2000, 1999, 1998 (All-time), 1998, 1997, 1996, 1995, 1994, 1993, 1991 (All-time), 1990 (All-time), 1989 (All-time))}
  \item{country}{Country of the artist ()}
}
}
\source{
\href{https://www.abc.net.au/triplej/hottest100/archive/search/}{ABC}
}
\usage{
historical_countdowns
}
\description{
A dataset of the Hottest 100 Countdown results for all available years.
This will be updated as each new Countdown is added in each consecutive year.
The variables include:
}
\references{
ABC (2020) The Hottest 100 Archive
(\href{https://www.abc.net.au/triplej/hottest100/archive/search/}{ABC})
}
\keyword{datasets}
