% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotteR.R
\docType{package}
\name{hotteR}
\alias{hotteR}
\title{Hottest 100 Countdown and Triple J Plays Webscraping and Analysis in R}
\description{
Hottest 100 Countdown and Triple J Plays Webscraping and Analysis in R
}
\details{
Hottest 100 Countdown and Triple J Analysis in R
}
