% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hotteR.R
\docType{data}
\name{hotteR_palette}
\alias{hotteR_palette}
\title{Builds the main function for styling ggplot2 plots in the hotteR style
Import the correct colours as hex codes to be used}
\format{
An object of class \code{character} of length 5.
}
\usage{
hotteR_palette
}
\description{
@author Trent Henderson
}
\keyword{datasets}
