% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior.R
\name{plot_prior}
\alias{plot_prior}
\title{Function to produce a density plot of the prior distribution of
nationality coefficients to visualise the shape before fitting
a probabilistic prediction model}
\usage{
plot_prior(timescale = c("Last Decade", "All Time"))
}
\arguments{
\item{timescale}{a value of either ["Last Decade", "All Time"] indicating the timescale to calculate over}
}
\value{
a density plot of class `ggplot` that shows the distribution of coefficient values in odds
}
\description{
Function to produce a density plot of the prior distribution of
nationality coefficients to visualise the shape before fitting
a probabilistic prediction model
}
\author{
Trent Henderson
}
