% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_probabilities.R
\name{plot_probabilities}
\alias{plot_probabilities}
\title{Function to produce density plots of probability of membership by quartile and nationality}
\usage{
plot_probabilities(
  data = historical_countdowns,
  timescale = c("Last Decade", "All Time")
)
}
\arguments{
\item{data}{The dataframe of Hottest 100 results to analyse}

\item{timescale}{a value of either "Last Decade" or "All Time" indicating the timescale to calculate over}
}
\value{
an object of class ggpot which holds the matrix of density plots
}
\description{
Function to produce density plots of probability of membership by quartile and nationality
}
\examples{
\dontrun{
plot_probabilities(timescale = "Last Decade")
}

}
\author{
Trent Henderson
}
