% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hotteR.R
\name{theme_hotteR}
\alias{theme_hotteR}
\title{Style general plot features according to the hotteR theme}
\usage{
theme_hotteR(grids = FALSE)
}
\description{
Style general plot features according to the hotteR theme
}
