% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Sets default geom colours to be compliant with The Runoff branding}
\usage{
.onAttach(libname, pkgname)
}
\description{
Sets default geom colours to be compliant with The Runoff branding
}
