% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countdowns.R
\name{get_countdowns}
\alias{get_countdowns}
\title{Function to webscrape historical Triple J Hottest 100 Countdown results
and return a dataframe of cleaned results in an analysis-ready format. Note
that this function is highly unstable due to the nature of RSelenium.}
\usage{
get_countdowns(sleep = 1)
}
\arguments{
\item{sleep}{an integer specifying the amount of computer sleep time befoere each RSelenium iteration}
}
\value{
a dataframe of results for all years in a tidy format
}
\description{
Function to webscrape historical Triple J Hottest 100 Countdown results
and return a dataframe of cleaned results in an analysis-ready format. Note
that this function is highly unstable due to the nature of RSelenium.
}
\examples{
\dontrun{
get_countdowns(sleep = 1)
}

}
\author{
Joy Horng
}
